win-requirements.txt
--extra-index-url https://kivy.org/downloads/packages/simple/
pypiwin32
kivy.deps.sdl2
kivy.deps.glew
kivy.deps.gstreamer